Fix/ filesystem permission

+++ app/Http/Controllers/PostController.php
@@ -6,7 +6,6 @@ use App\Http\Requests\StorePostsRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Models\Category;
 use App\Models\Post;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
@@ -23,13 +22,13 @@ class PostController extends Controller
 
         if (Gate::check('post_access_all')) {
             return inertia('Post/Index', [
-                'posts' => Post::with('user', 'categories')->get()
+                'posts' => Post::with('user', 'categories')->latest()->get()
             ]);
         }
 
         if (Gate::check('post_access_self')) {
             return inertia('Post/Index', [
-                'posts' => Post::with('user', 'categories')->where('user_id', auth()->user()->id)->get()
+                'posts' => Post::with('user', 'categories')->where('user_id', auth()->user()->id)->latest()->get()
             ]);
         }
     }
@@ -77,12 +76,15 @@ class PostController extends Controller
     public function show(Post $post)
     {
         Gate::authorize('post_show');
+        if (Gate::allows('post_access_all') || $post->user_id === auth()->id()) {
+            $post = $post->load('user', 'categories');
+            $post->featured_image = $post->featured_image ? asset("storage/posts/{$post->featured_image}") : null;
+            return inertia('Post/View', [
+                'post' => $post
+            ]);
+        }
 
-        $post = $post->load('user', 'categories');
-        $post->featured_image = $post->featured_image ? asset("storage/posts/{$post->featured_image}") : null;
-        return inertia('Post/View', [
-            'post' => $post
-        ]);
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -91,13 +93,16 @@ class PostController extends Controller
     public function edit(Post $post)
     {
         Gate::authorize('post_edit');
-        $categories = Category::all();
-        $post = $post->load('categories');
-        $post->featured_image = $post->featured_image ? asset("storage/posts/{$post->featured_image}") : null;
-        return inertia('Post/Edit', [
-            'post' => $post->load('categories'),
-            'categories' => $categories,
-        ]);
+        if (Gate::allows('post_access_all') || $post->user_id === auth()->id()) {
+            $categories = Category::all();
+            $post = $post->load('categories');
+            $post->featured_image = $post->featured_image ? asset("storage/posts/{$post->featured_image}") : null;
+            return inertia('Post/Edit', [
+                'post' => $post->load('categories'),
+                'categories' => $categories,
+            ]);
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -106,32 +111,34 @@ class PostController extends Controller
     public function update(UpdatePostRequest $request, Post $post)
     {
         Gate::authorize('post_edit');
+        if (Gate::allows('post_access_all') || $post->user_id === auth()->id()) {
+            if ($request->hasFile('featured_image')) {
+                if ($post->featured_image) {
+                    Storage::delete("public/posts/{$post->featured_image}");
+                }
 
-        if ($request->hasFile('featured_image')) {
-            if ($post->featured_image) {
-                Storage::delete("public/posts/{$post->featured_image}");
+                $featured_image = $request->file('featured_image');
+                $filename = $request->slug . '.' . $featured_image->getClientOriginalExtension();
+                $featured_image->storeAs('public/posts', $filename);
+                $post->featured_image = $filename;
+                $post->save();
             }
 
-            $featured_image = $request->file('featured_image');
-            $filename = $request->slug . '.' . $featured_image->getClientOriginalExtension();
-            $featured_image->storeAs('public/posts', $filename);
-            $post->featured_image = $filename;
-            $post->save();
+            if ($post->featured_image && ($request->slug !== $post->slug)) {
+                $oldFilename = $post->featured_image;
+                $newFilename = $request->slug . '.' . pathinfo($oldFilename, PATHINFO_EXTENSION);
+                Storage::move("public/posts/{$oldFilename}", "public/posts/{$newFilename}");
+                $post->featured_image = $newFilename;
+                $post->save();
+            }
+
+            $post->update($request->except('featured_image'));
+            $categories = $request->categories;
+            $post->categories()->sync($categories);
+
+            return redirect()->route('post.index')->with('success', 'Post updated successfully');
         }
-
-        if ($post->featured_image && ($request->slug !== $post->slug)) {
-            $oldFilename = $post->featured_image;
-            $newFilename = $request->slug . '.' . pathinfo($oldFilename, PATHINFO_EXTENSION);
-            Storage::move("public/posts/{$oldFilename}", "public/posts/{$newFilename}");
-            $post->featured_image = $newFilename;
-            $post->save();
-        }
-
-        $post->update($request->except('featured_image'));
-        $categories = $request->categories;
-        $post->categories()->sync($categories);
-
-        return redirect()->route('post.index')->with('success', 'Post updated successfully');
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -140,10 +147,12 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         Gate::authorize('post_delete');
-        if ($post->featured_image) {
-            Storage::delete("public/posts/{$post->featured_image}");
+        if (Gate::allows('post_access_all') || $post->user_id === auth()->id()) {
+            if ($post->featured_image) {
+                Storage::delete("public/posts/{$post->featured_image}");
+            }
+            $post->delete();
+            return redirect()->route('post.index')->with('success', 'Post deleted successfully');
         }
-        $post->delete();
-        return redirect()->route('post.index')->with('success', 'Post deleted successfully');
     }
 }
